# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "BookStore APIs exposed from a Java EE back-end to an Angular front-end",
    "version" : "1.0.0",
    "title" : "BookStore APIs",
    "contact" : {
      "name" : "Moises Gil",
      "email" : "moisesgilsantana@gmail.com"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/bookstore-back/api",
  "tags" : [ {
    "name" : "Book"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/books" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Bring all the books registered",
        "description" : "",
        "operationId" : "getAllBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Book" ],
        "summary" : "Create a book with a given JSON representation of the book",
        "description" : "",
        "operationId" : "createBook",
        "consumes" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "201" : {
            "description" : "Book Created"
          },
          "500" : {
            "description" : "Error"
          }
        }
      }
    },
    "/books/count" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns the quantity of books registered in the system",
        "description" : "",
        "operationId" : "countBooks",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Find a book with a given Book ID",
        "description" : "",
        "operationId" : "getBook",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Book ID",
          "required" : true,
          "type" : "integer",
          "minimum" : 1.0,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Book Found",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "404" : {
            "description" : "Book not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Book" ],
        "summary" : "Delete a book with a given ID",
        "description" : "",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Book deleted",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Book" : {
      "type" : "object",
      "required" : [ "description", "imageUrl", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 10000
        },
        "language" : {
          "type" : "string",
          "enum" : [ "ENGLISH", "FRENCH", "SPANISH", "PORTUGUESE", "ITALIAN", "FINISH", "GERMAN", "DEUTSCH", "RUSSIAN" ]
        },
        "unitCost" : {
          "type" : "number",
          "format" : "float",
          "minimum" : 1.0
        },
        "isbn" : {
          "type" : "string",
          "description" : "ISBN Number",
          "minLength" : 1,
          "maxLength" : 50
        },
        "publicationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "nbOfPages" : {
          "type" : "integer",
          "format" : "int32",
          "minimum" : 1.0
        },
        "imageUrl" : {
          "type" : "string",
          "description" : "URL of the image cover"
        }
      }
    }
  }
}